Register event handlers on div.auth instead of document.body
As per Antony's advice, register the event handlers to a div that exists in index.html on load instead of registering them to the document.body.

This means we need to keep the '.auth' div in index.html (or make sure to change these event handlers if we change the structure of index.html).