Create Ui functionality for Show All Events
Created a handlebars template to be invoked in the allEventsSuccess
function in the ui file. This shows all the events created by any user
and are wrapped in a link so that in the future we'll be able
to click on the event 'box' it will go to the single event view.