Create Show for Single Event View
Added another GET request to show a single event. This is called when
the event link in All Events view is clicked on and renders a new view
that shows all of the info for that single event.

We've added edit and delete buttons to the single-event handlebars
template, but those still need to be 'activated'.