Make start and end time options functional
Add functionality for converting the date and time formats between what Mongo expects and what humans like to see.

Use this functionality when sending a new event or an updated event to the server.

Also use it when getting data back from the server, so we can display a single event better.

Code is heavily commented and mostly uses JavaScript string manipulation & template literals to format stuff.