Add My Events functionality
added ui functionality to take a "my events" template to display my events
change "all events" to use a partial to be shared with my events because they
display the same thing
rename div that holds all events to "events list" so we can use it for both.
add my-events template to share with all events.