Add RSVP view to Single Event View
When a user is logged in and clicks on another user's event the view
state changes to show the event info + the RSVP info which is:

Are you coming?
Yes, No, Maybe

We created a new handlebars template to render this rsvp view.

Next, we'll add a submit rsvp button to send the ajax request to post a
new rsvp to that event.