Add edit RSVP functionality
Clicking on the "edit" link in the single RSVP view calls a function in events/events.js.
That function displays the "single Event view," which is either the RSVP form or the event view.

If you don't own the event, you'll see the RSVP form.

Submitting that form updates your RSVP using the rsvps#createorupdate method in the RSVP controller.