Display Rsvp Data on a Single Event
Added code to the Ui function for singleEventSuccess to count
the number of each type of RSVP.

When a user clicks on an event it will show how many people have
responded yes, no or maybe.

Still need to think about whether this data should also be
viewable from the my Rsvps view.