Flexibly calculate response totals in events/ui.js
Declare an empty responses object.

Loop through all of the questions and add a new property to the responses object that is named after the question text.

Loop through all of the possible answer options and set up a new object with a single key (the answer option) and a value of 0.

The resulting responses object looks like this:

```js
responses: {
  'Are you coming?': [
    { 'Yes': 0 },
    { 'No': 0 },
    { 'Maybe': 0 },
  ],
}
```

Loop through each RSVP and update the response count for the appropriate answer option.

Attach the new responses object to the event data as event.responses before passing that data to the Handlebars templates.