Create HTML Structure - Interface & Homepage
Created starter html to structure the profile view and the homepage
view.

Profile view includes Title, nav, interface and failure/message divs.
Nav has a class of hide by default for now.

Created a handlebars template for the homepage that loads on doc
ready to display the title and sign up/sign buttons.