Add prompts for users who have no RSVPs or no events
Add conditional messages inside of the handlebars templates for my events and my RSVPs.

Messages display if user has no RSVPs or no events.
Messages include links to create event and find an event to which they can RSVP.

(To make these links work, we need to use classes instead of IDs in the links in the nav.)